+++ src/components/composite/sections/ContextSection.tsx
@@ -3,24 +3,23 @@ import PeopleAltIcon from '@mui/icons-material/PeopleAlt';
 import PublicIcon from '@mui/icons-material/Public';
 import SchoolIcon from '@mui/icons-material/School';
 import { Box, Grid, Typography } from '@mui/material';
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { ReactElement } from 'react-markdown';
-import { AspectCardFragment, Reference } from '../../../models/graphql-schema';
-import { CardLayoutContainer, CardLayoutItem } from '../../core/CardLayoutContainer/CardLayoutContainer';
+import { Reference } from '../../../models/graphql-schema';
+// import { CardLayoutContainer, CardLayoutItem } from '../../core/CardLayoutContainer/CardLayoutContainer';
 import Markdown from '../../core/Markdown';
 import { SectionSpacer } from '../../core/Section/Section';
 import SectionHeader from '../../core/Section/SectionHeader';
-import MembershipBackdrop from '../common/Backdrops/MembershipBackdrop';
-import AspectCard from '../common/cards/AspectCard/AspectCard';
+// import MembershipBackdrop from '../common/Backdrops/MembershipBackdrop';
+// import AspectCard from '../common/cards/AspectCard/AspectCard';
 import DashboardGenericSection from '../common/sections/DashboardGenericSection';
 import TagsComponent from '../common/TagsComponent/TagsComponent';
-import Button from '@mui/material/Button';
-import AspectCreationDialog, { AspectCreationOutput } from '../aspect/AspectCreationDialog/AspectCreationDialog';
-import { AspectCardFragmentDoc, useCreateAspectMutation } from '../../../hooks/generated/graphql';
-import { useApolloErrorHandler, useNotification, useUrlParams } from '../../../hooks';
-import CardFilter from '../../core/card-filter/CardFilter';
-import { aspectTagsValueGetter, aspectValueGetter } from '../../core/card-filter/value-getters/aspect-value-getter';
+// import Button from '@mui/material/Button';
+// import AspectCreationDialog, { AspectCreationOutput } from '../aspect/AspectCreationDialog/AspectCreationDialog';
+// import { AspectCardFragmentDoc, useCreateAspectMutation } from '../../../hooks/generated/graphql';
+// import CardFilter from '../../core/card-filter/CardFilter';
+// import { aspectTagsValueGetter, aspectValueGetter } from '../../core/card-filter/value-getters/aspect-value-getter';
 import References from '../common/References/References';
 
 export interface ContextSectionProps {
@@ -35,14 +34,14 @@ export interface ContextSectionProps {
   impact?: string;
   who?: string;
   references?: Reference[];
-  aspects?: AspectCardFragment[];
+  // aspects?: AspectCardFragment[];
   aspectsLoading?: boolean;
   canReadAspects?: boolean;
   canCreateAspects?: boolean;
 }
 
 const ContextSection: FC<ContextSectionProps> = ({
-  contextId,
+  // contextId,
   primaryAction,
   banner,
   background,
@@ -53,98 +52,12 @@ const ContextSection: FC<ContextSectionProps> = ({
   impact,
   who,
   references,
-  aspects = [],
-  aspectsLoading,
-  canReadAspects,
-  canCreateAspects,
+  // aspects = [],
+  // aspectsLoading,
+  // canReadAspects,
+  // canCreateAspects,
 }) => {
   const { t } = useTranslation();
-  const handleError = useApolloErrorHandler();
-  const notify = useNotification();
-  const { hubNameId = '', challengeNameId = '', opportunityNameId = '' } = useUrlParams();
-  const [aspectDialogOpen, setAspectDialogOpen] = useState(false);
-
-  // todo: move handlers to the contextTabContainer
-  const [createAspect] = useCreateAspectMutation({
-    onError: handleError,
-    onCompleted: () => notify(t('components.context-section.create-aspect'), 'success'),
-    update: (cache, { data }) => {
-      if (!data) {
-        return;
-      }
-      const { createAspectOnContext } = data;
-
-      const contextRefId = cache.identify({
-        __typename: 'Context',
-        id: contextId,
-      });
-
-      if (!contextRefId) {
-        return;
-      }
-
-      cache.modify({
-        id: contextRefId,
-        fields: {
-          aspects(existingAspects = []) {
-            const newAspectRef = cache.writeFragment({
-              data: createAspectOnContext,
-              fragment: AspectCardFragmentDoc,
-              fragmentName: 'AspectCard',
-            });
-            return [...existingAspects, newAspectRef];
-          },
-        },
-      });
-    },
-  });
-
-  const handleCreateDialogOpened = () => setAspectDialogOpen(true);
-  const handleCreateDialogClosed = () => setAspectDialogOpen(false);
-  const onCreate = async (aspect: AspectCreationOutput) => {
-    if (!contextId) {
-      return;
-    }
-
-    createAspect({
-      variables: {
-        aspectData: {
-          contextID: contextId,
-          displayName: aspect.displayName,
-          description: aspect.description,
-          type: aspect.type,
-          tags: aspect.tags,
-        },
-      },
-      optimisticResponse: {
-        createAspectOnContext: {
-          __typename: 'Aspect',
-          id: '',
-          nameID: aspect.nameID || '',
-          displayName: aspect.displayName ?? '',
-          description: aspect.description,
-          type: aspect.type,
-          tagset: {
-            id: '-1',
-            name: 'default',
-            tags: aspect.tags ?? [],
-          },
-          banner: {
-            id: '-1',
-            name: '',
-            uri: '',
-          },
-          bannerNarrow: {
-            id: '-1',
-            name: '',
-            uri: '',
-          },
-        },
-      },
-    });
-
-    setAspectDialogOpen(false);
-  };
 
   return (
     <>
@@ -216,54 +129,6 @@ const ContextSection: FC<ContextSectionProps> = ({
           >
             <Typography component={Markdown} variant="body1" children={who} />
           </DashboardGenericSection>
-        </Grid>
-        <Grid item xs={12}>
-          <MembershipBackdrop show={!canReadAspects} blockName={t('common.aspects')}>
-            <DashboardGenericSection
-              headerText={`${t('common.aspects')} (${aspects ? aspects.length : 0})`}
-              primaryAction={
-                canCreateAspects && (
-                  <Button variant="contained" onClick={handleCreateDialogOpened}>
-                    {t('buttons.create')}
-                  </Button>
-                )
-              }
-            >
-              {aspectsLoading ? (
-                <CardLayoutContainer>
-                  <CardLayoutItem>
-                    <AspectCard loading={true} />
-                  </CardLayoutItem>
-                  <CardLayoutItem>
-                    <AspectCard loading={true} />
-                  </CardLayoutItem>
-                </CardLayoutContainer>
-              ) : (
-                <CardFilter data={aspects} tagsValueGetter={aspectTagsValueGetter} valueGetter={aspectValueGetter}>
-                  {filteredAspects => (
-                    <CardLayoutContainer>
-                      {filteredAspects.map((x, i) => (
-                        <CardLayoutItem key={i}>
-                          <AspectCard
-                            aspect={x}
-                            hubNameId={hubNameId}
-                            challengeNameId={challengeNameId}
-                            opportunityNameId={opportunityNameId}
-                          />
-                        </CardLayoutItem>
-                      ))}
-                    </CardLayoutContainer>
-                  )}
-                </CardFilter>
-              )}
-            </DashboardGenericSection>
-          </MembershipBackdrop>
-          <AspectCreationDialog
-            open={aspectDialogOpen}
-            onCancel={handleCreateDialogClosed}
-            onCreate={onCreate}
-            aspectNames={aspects.map(x => x.displayName)}
-          />
         </Grid>
       </Grid>
     </>
