+++ src/views/Hub/HubContextView.tsx
@@ -50,7 +50,7 @@ export const HubContextView: FC<HubContextViewProps> = ({ entities, state, optio
     id = '',
   } = context || ({} as Context);
   const hubBanner = getVisualBanner(visuals);
-  const aspects = entities?.aspects;
+  // const aspects = entities?.aspects;
   const references = entities?.references;
 
   return (
@@ -78,7 +78,6 @@ export const HubContextView: FC<HubContextViewProps> = ({ entities, state, optio
       who={who}
       contextId={id}
       references={references}
-      aspects={aspects}
       aspectsLoading={loading}
       canReadAspects={canReadAspects}
       canCreateAspects={canCreateAspects}
